apps/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-socketio==5.13.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.4
python-jose[cryptography]==3.5.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
httpx==0.28.1
